requirements.txt
google-adk>=1.27.0
google-genai>=1.67.0
fastapi>=0.135.0
uvicorn>=0.41.0
python-dotenv>=1.2.0
websockets>=15.0.0
pillow>=12.0.0
